# internal/appview/openapi.yaml
openapi: 3.0.3
info:
  title: HashPost AppView API
  description: HashPost AppView provides a unified API for accessing HashPost data aggregated from the PDS
  version: 1.0.0
  contact:
    name: HashPost Team
    email: team@hashpost.com
servers:
  - url: http://localhost:8081
    description: Development server
  - url: https://appview.hashpost.com
    description: Production server
paths:
...
  /api/v1/auth/login:
    post:
      summary: Login user
      description: Authenticate user with email/password
      operationId: loginUser
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                password:
                  type: string
                  description: User's password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/register:
    post:
      summary: Register user
      description: Create a new user account
      operationId: registerUser
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - handle
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                password:
                  type: string
                  description: User's password
                handle:
                  type: string
                  description: User's handle
                inviteCode:
                  type: string
                  description: Optional invite code
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRegistrationResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Handle already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/me:
    get:
      summary: Get current user session
      description: Get information about the currently authenticated user
      operationId: getCurrentUserSession
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User session information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSession'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/logout:
    post:
      summary: Logout user
      description: Logout the current user session
      operationId: logoutUser
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logged out successfully"
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    UserLoginResponse:
      type: object
      required:
        - accessToken
        - refreshToken
        - handle
        - did
        - email
      properties:
        accessToken:
          type: string
          description: JWT access token
        refreshToken:
          type: string
          description: JWT refresh token
        handle:
          type: string
          description: User's handle
        did:
          type: string
          description: User's DID
        email:
          type: string
          format: email
          description: User's email address
    UserRegistrationResponse:
      type: object
      required:
        - accessToken
        - refreshToken
        - handle
        - did
        - email
      properties:
        accessToken:
          type: string
          description: JWT access token
        refreshToken:
          type: string
          description: JWT refresh token
        handle:
          type: string
          description: User's handle
        did:
          type: string
          description: User's DID
        email:
          type: string
          format: email
          description: User's email address
    UserSession:
      type: object
      required:
        - handle
        - did
        - email
      properties:
        handle:
          type: string
          description: User's handle
        did:
          type: string
          description: User's DID
        email:
          type: string
          format: email
          description: User's email address
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details

    # RBAC Schemas
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from PDS authentication
...