# internal/database/generated/pds/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package generated

import (
	"time"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Comment struct {
	ID         uuid.UUID          `json:"id"`
	UserID     pgtype.UUID        `json:"user_id"`
	PostID     pgtype.UUID        `json:"post_id"`
	ParentID   pgtype.UUID        `json:"parent_id"`
	Content    string             `json:"content"`
	AtprotoUri *string            `json:"atproto_uri"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
	UpdatedAt  pgtype.Timestamptz `json:"updated_at"`
}

type DpopNonce struct {
	ID        uuid.UUID          `json:"id"`
	Nonce     string             `json:"nonce"`
	ExpiresAt time.Time          `json:"expires_at"`
	UsedAt    pgtype.Timestamptz `json:"used_at"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

// OAuth client registrations with external PDS servers
type ExternalPdsClient struct {
	ID           uuid.UUID          `json:"id"`
	PdsEndpoint  string             `json:"pds_endpoint"`
	ClientID     string             `json:"client_id"`
	ClientSecret *string            `json:"client_secret"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
	UpdatedAt    pgtype.Timestamptz `json:"updated_at"`
}

type OauthAccessToken struct {
	ID           uuid.UUID          `json:"id"`
	AccessToken  string             `json:"access_token"`
	RefreshToken *string            `json:"refresh_token"`
	ClientID     string             `json:"client_id"`
	UserDid      string             `json:"user_did"`
	Scope        string             `json:"scope"`
	ExpiresAt    time.Time          `json:"expires_at"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
}

type OauthAuthorizationCode struct {
	ID          uuid.UUID          `json:"id"`
	Code        string             `json:"code"`
	ClientID    string             `json:"client_id"`
	UserDid     string             `json:"user_did"`
	RedirectUri string             `json:"redirect_uri"`
	Scope       string             `json:"scope"`
	Nonce       *string            `json:"nonce"`
	ExpiresAt   time.Time          `json:"expires_at"`
	UsedAt      pgtype.Timestamptz `json:"used_at"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
}

type OauthClient struct {
	ID            uuid.UUID          `json:"id"`
	ClientID      string             `json:"client_id"`
	ClientName    string             `json:"client_name"`
	RedirectUris  []string           `json:"redirect_uris"`
	Scopes        []string           `json:"scopes"`
	GrantTypes    []string           `json:"grant_types"`
	ResponseTypes []string           `json:"response_types"`
	CreatedAt     pgtype.Timestamptz `json:"created_at"`
	UpdatedAt     pgtype.Timestamptz `json:"updated_at"`
}

type Post struct {
	ID         uuid.UUID          `json:"id"`
	UserID     pgtype.UUID        `json:"user_id"`
	SubforumID pgtype.UUID        `json:"subforum_id"`
	Title      string             `json:"title"`
	Content    string             `json:"content"`
	AtprotoUri *string            `json:"atproto_uri"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
	UpdatedAt  pgtype.Timestamptz `json:"updated_at"`
}

type Subforum struct {
	ID          uuid.UUID          `json:"id"`
	Name        string             `json:"name"`
	Slug        string             `json:"slug"`
	Description *string            `json:"description"`
	CreatedBy   pgtype.UUID        `json:"created_by"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
	UpdatedAt   pgtype.Timestamptz `json:"updated_at"`
	AtprotoUri  *string            `json:"atproto_uri"`
	PrefixType  string             `json:"prefix_type"`
}

type SubforumSubscription struct {
	ID         uuid.UUID          `json:"id"`
	UserID     pgtype.UUID        `json:"user_id"`
	SubforumID pgtype.UUID        `json:"subforum_id"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
}

type User struct {
	ID        uuid.UUID          `json:"id"`
	Did       string             `json:"did"`
	Handle    string             `json:"handle"`
	Email     *string            `json:"email"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
	// bcrypt hashed password for user authentication
	PasswordHash *string `json:"password_hash"`
}

type UserSession struct {
	ID             uuid.UUID          `json:"id"`
	SessionID      string             `json:"session_id"`
	UserDid        string             `json:"user_did"`
	Handle         string             `json:"handle"`
	CreatedAt      pgtype.Timestamptz `json:"created_at"`
	ExpiresAt      time.Time          `json:"expires_at"`
	LastAccessedAt pgtype.Timestamptz `json:"last_accessed_at"`
}

type Vote struct {
	ID        uuid.UUID          `json:"id"`
	UserID    pgtype.UUID        `json:"user_id"`
	PostID    pgtype.UUID        `json:"post_id"`
	CommentID pgtype.UUID        `json:"comment_id"`
	VoteType  string             `json:"vote_type"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}